refactor(BUX-686): splitting to miltiple files
